API_s/requirements.txt
scikit-learn==1.4.2
fastapi<1
uvicorn<1
requests
pandas
python-jose
dotenv